# Defer slow imports of matplotlib, networkx and sympy.