Old:
```bash
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  1.50s user 0.24s system 138% cpu 1.251 total
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  1.49s user 0.24s system 138% cpu 1.242 total
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  1.49s user 0.24s system 138% cpu 1.253 total
```

New:
```bash
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  0.42s user 0.12s system 239% cpu 0.223 total
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  0.38s user 0.11s system 243% cpu 0.204 total
(venv)  > time python -c 'import discopy'
python -c 'import discopy'  0.38s user 0.11s system 241% cpu 0.203 total
```

Average 1.25 / 0.21 ≈ 6x speed-up