@y-richie-y 2022-09-14
```
> time python -c 'import discopy'
python -c 'import discopy'  2.34s user 0.67s system 88% cpu 3.392 total
> time python -c 'import discopy'
python -c 'import discopy'  2.10s user 0.25s system 124% cpu 1.895 total
> time python -c 'import discopy'
python -c 'import discopy'  2.03s user 0.22s system 127% cpu 1.762 total

> time python -c 'import discopy'
python -c 'import discopy'  0.45s user 0.10s system 170% cpu 0.319 total
> time python -c 'import discopy'
python -c 'import discopy'  0.43s user 0.11s system 174% cpu 0.304 total
> time python -c 'import discopy'
python -c 'import discopy'  0.41s user 0.09s system 174% cpu 0.287 total
```

Even bigger speedup on my machine. Nice!